U/ben42code/fix readme (#8)
* Fix README

README.md file is used as the main page for github but also for pypi.
Keep it self contained and simple...with no broken links 😅

* Fix README

README.md file is used as the main page for github but also for pypi.
Keep it self contained and simple...with no broken links 😅